🎨 Style : header 폴더 위치 components 안으로 변경
